Fix stars
Instead of checking new stargazers, simply overwrite the stars.csv
file. Unlike the views and clones stats, stargazers data is always
available, so we do not need to store the data in the csv file and
read it before adding new data. This solves a problem that we can
have when a user change its name and the csv file repeats that star,
since the code was seeing that as a 'new' stargazer.