Revert upload after modification to stats_dict
This revert the commit that placed the upload after the modification
to the stats_dict. The csv uploaded to the s3 bucket has a different
time format than that used by bokeh in the returned stats_dict,
that's why, for now, the upload of the csv file needs to be done
before altering the stats_dict.